+++ src/Products/PloneMeeting/tests/testMeetingItem.py
@@ -9196,22 +9196,38 @@ class testMeetingItem(PloneMeetingTestCase):
         self.assertTrue(item.show_field('groupsInChargeNotes'))
         # even if item editable, field can be not editable if condition if False
         self.assertFalse(item.mayQuickEdit('groupsInChargeNotes'))
+        # bypass for Manager
+        self.changeUser('siteadmin')
+        self.assertTrue(item.show_field('groupsInChargeNotes')
+                        and item.mayQuickEdit('groupsInChargeNotes'))
         # group in charge can view and edit
         self.changeUser('pmObserver2')
         self.assertTrue(self.hasPermission(View, item))
         self.assertTrue(item.show_field('groupsInChargeNotes'))
         self.assertTrue(item.mayQuickEdit('groupsInChargeNotes'))
+        # bypass for Manager
+        self.changeUser('siteadmin')
+        self.assertTrue(item.show_field('groupsInChargeNotes')
+                        and item.mayQuickEdit('groupsInChargeNotes'))
         # with no group in charge
         self.changeUser('pmCreator1')
         item.setGroupsInCharge([])
         self.assertTrue(item.show_field('groupsInChargeNotes'))
         self.assertFalse(item.mayQuickEdit('groupsInChargeNotes'))
+        # bypass for Manager
+        self.changeUser('siteadmin')
+        self.assertTrue(item.show_field('groupsInChargeNotes')
+                        and item.mayQuickEdit('groupsInChargeNotes'))
         # make proposing group only able to edit
         self.changeUser('pmObserver2')
         self._setupItemFieldsConfig(
             'groupsInChargeNotes',
             edit='python: tool.user_is_in_org(org_uid=item.getProposingGroup())')
         self.assertFalse(item.mayQuickEdit('groupsInChargeNotes'))
+        # bypass for Manager
+        self.changeUser('siteadmin')
+        self.assertTrue(item.show_field('groupsInChargeNotes')
+                        and item.mayQuickEdit('groupsInChargeNotes'))
         self.changeUser('pmCreator1')
         self.assertTrue(item.mayQuickEdit('groupsInChargeNotes'))
         # wrong condition, will raise if used
@@ -9219,6 +9235,10 @@ class testMeetingItem(PloneMeetingTestCase):
         self.assertRaises(NameError, item.mayQuickEdit, 'groupsInChargeNotes')
         # does not raise if not used
         self._enableField(['groupsInChargeNotes'], enable=False)
+        self.assertFalse(item.mayQuickEdit('groupsInChargeNotes'))
+        # bypass for Manager not working if field not enabled
+        self.changeUser('siteadmin')
+        self.assertFalse(item.show_field('groupsInChargeNotes'))
         self.assertFalse(item.mayQuickEdit('groupsInChargeNotes'))